pyproject.toml
[tool.poetry]
name = "linux-playground"
version = "0.1.0"
description = "Linux input simulation tools"
authors = ["Eden Refael edenraf@hotmail.com"]
readme = "README.md"
packages = [{include = "linux_playground"}]

[tool.poetry.dependencies]
python = "^3.9"
evdev = "^1.6.1"
python-dotenv = "^1.1.1"
requests = "^2.32.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
